pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ingresarios_options_research"
version = "0.1.0"
description = "Scientific validation pipeline for SPX 0DTE Iron Condor strategies"
authors = [{name = "Antigravity", email = "antigravity@example.com"}]
dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "pyyaml",
    "polars", 
    "pyarrow",
    "matplotlib",
    "seaborn",
    "tqdm"
]
requires-python = ">=3.9"

[tool.setuptools.packages.find]
where = ["src"]
